pyproject.toml
[tool.poetry]
name = "vespa-plugin"
version = "0.1.0"
description = ""
authors = ["Simon Lee <simon2020291@gmail.com>"]
readme = "README.md"
packages = [{include = "vespa_plugin"}]

[tool.poetry.dependencies]
python = "~3.10"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
pinecone-client = "^2.2.1"
tenacity = "^8.2.2"
arrow = "^1.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
